feat: Add blog title, description and sidebar count
This commit adds a blog title, description and sidebar count to the Docusaurus configuration file. The blog title is set to "4lch4's Blog" and the description is set to "A blog about my life in the world of software and anything I find interesting." The sidebar count is set to 'ALL'.